+++ manual/Changelog.md
@@ -1,4 +1,8 @@
 
+###v0.2.8
+- expanded the analysed elements **FSnode::parse_url()** can give; filename, directory, archive, database, table, namespace, resource, separator, masterdivider, divider, assigner. And allows for returning a selection of those elements.
+- adds **FSnode::mime_content_type()** method to the default class diagram
+
 ###v0.2.7
 - added detection in **FSnode::parse_url()** for files inside compressed archives.
 - have been wondering about how **[parse_url()](http://php.net/parse_url)** maps the [URI_scheme](http://en.wikipedia.org/wiki/URI_scheme#Examples). Desided **FSnode::parse_url()** will have to return more *component* types: authority, directory, filename, extension, server (scheme://host:port), *(array)* query. In case of a compressed archive, the archive will be the directory and the internal path noted as filename.